requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0
django-cors-headers==3.2.0
django-heroku==0.3.1
django-tastypie==0.14.2
djangorestframework==3.10.3
gunicorn==20.0.4
psycopg2==2.8.4
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==4.1.4
